+++ HISTORY.md
@@ -1,6 +1,7 @@
 ## v8.0.1
 
 * [ADDED] Multi table update support for `mysql` and `postgres` [#60](https://github.com/doug-martin/goqu/issues/60)
+* [ADDED] `goqu.V` so values can be used on the LHS of expressions [#104](https://github.com/doug-martin/goqu/issues/104)
 
 ## v8.0.0